Custom Packets Setup
Created a new namespace to store the Custom Variant Settings.

Created a new Custom Packet for sending the Custom Variant Settings to the clients. Stored all functions related inside the Custom Variants Settings.

Added a new Util function isFloat to test if a string would properly convert to a float.

 First fully working Custom Variant Setting Gravity implimented.