Initial Push
A combined pushed of all the changes for json config

Created a new class that allows for easier interaction with a RapidJson document called easy_json. Allows you to interact with the json document in a more intuitive fashion handling all of the allocations and conversions internally. Also allows you to add specializations for custom data types (i.e real_point3d)

using the new easy_json class converted the config file over to a json format. Taking into account that the old config file was a ini configuration a method for upgrading to the json file format has been added.