More logging and cleanup
more logging to help with any form of errors that happen during usage.

removed the defaultvalues map from the class instead relying on the default value to be passed through when attempting to get a value.

removed the old easy_json and replaced it with easy_json_struct renaming that to just easy_json