error handling on saving json
easy_json will detect failed saving and determine which key value pair is causing the issue. Then it will attempt removing that key from the json document and resaving. It will do this until it is successful or ultimately fails, in which case it will then reset the json file to an empty json key and on next launch the game will load with default settings.