Loading error handling and ImGui changes
Changed ImGui to initialize the window list prior to initialize being called.. want to do this better but this method works.

Heavily modified ImGui Message box to give the option of defining multiple option choices with callback functions to execute depending on what the user chooses.

added an accessable member to the easy_json class that will give you the last error_code and an error message. this is being utilized by Config to display the error message in a MessageBox call that tells the user the configuration failed to load and tells them they have to fix the config or reset it.